css: use 64-bit selector expansion weights and charge numeric tokens
Selector weight estimation truncated byte lengths to u32, so payloads
past 4 GB could wrap to a small weight instead of charging the budget.
Keep the whole weight chain in u64 (usize to u64 is lossless on every
supported target). Charge numeric tokens with a constant bound derived
from dtoa's shortest round-trip form instead of the generic flat rate.

+++ src/css/rules/mod.rs
@@ -1263,7 +1263,7 @@ pub struct MinifyContext<'a, 'bump> {
     /// average selector weight of their lists. `0` at the top level;
     /// maintained alongside `selector_expansion_multiplier` in
     /// `StyleRule::minify_nested_rules`.
-    pub(crate) selector_expansion_chain_bytes: u32,
+    pub(crate) selector_expansion_chain_bytes: u64,
     /// Running estimate of the serialized bytes the expansion will produce,
     /// checked against [`MAX_SELECTOR_EXPANSION_BYTES`].
     pub(crate) selector_expansion_bytes_total: u64,
